Inline CurrentPlaylistView
